Finish debug multiple expression evaluation
The calculator can now evaluate one single pair of number for each calculation accurately. There was an error in HTML's display for multiplication sign: it uses "&times" which really is a multiplication sign as checked in the ASCII table, while multiply function takes "x" (which has a different ASCII code) as a parameter.